@coffeegoddd 2024-11-14
@max-hoffman __DOLT__
 |      test_name      | from_latency_p95 | to_latency_p95 | percent_change |
|---------------------|------------------|----------------|----------------|
| tpcc-scale-factor-1 | 61.08            | 59.99          | -1.78          |


|      test_name      | from_server_name |           from_server_version            | from_tps | to_server_name |            to_server_version             | to_tps | percent_change |
|---------------------|------------------|------------------------------------------|----------|----------------|------------------------------------------|--------|----------------|
| tpcc-scale-factor-1 | dolt             | 0e34d26c17d8438db614724429d522e3e121e91e | 40.64    | dolt           | 4fa6227c36f54b0f103f679684262de91cf5f67f | 40.7   | 0.15           |