@coffeegoddd 2024-11-15
@max-hoffman __DOLT__
 |      read_tests      | from_latency | to_latency | percent_change |
|----------------------|--------------|------------|----------------|
| covering_index_scan  | 0.62         | 0.62       | 0.0            |
| groupby_scan         | 16.41        | 16.41      | 0.0            |
| index_join           | 2.26         | 2.26       | 0.0            |
| index_join_scan      | 1.79         | 1.64       | -8.38          |
| index_scan           | 53.85        | 55.82      | 3.66           |
| oltp_point_select    | 0.26         | 0.27       | 3.85           |
| oltp_read_only       | 5.28         | 5.37       | 1.7            |
| select_random_points | 0.64         | 0.65       | 1.56           |
| select_random_ranges | 0.63         | 0.64       | 1.59           |
| table_scan           | 54.83        | 55.82      | 1.81           |
| types_table_scan     | 139.85       | 144.97     | 3.66           |


|      write_tests      | from_latency | to_latency | percent_change |
|-----------------------|--------------|------------|----------------|
| oltp_delete_insert    | 5.77         | 5.88       | 1.91           |
| oltp_insert           | 2.91         | 2.91       | 0.0            |
| oltp_read_write       | 11.24        | 11.45      | 1.87           |
| oltp_update_index     | 2.91         | 2.97       | 2.06           |
| oltp_update_non_index | 2.86         | 2.91       | 1.75           |
| oltp_write_only       | 5.88         | 5.99       | 1.87           |
| types_delete_insert   | 6.21         | 6.21       | 0.0            |