@coffeegoddd 2024-11-19
@max-hoffman __DOLT__
 |      test_name      | from_latency_p95 | to_latency_p95 | percent_change |
|---------------------|------------------|----------------|----------------|
| tpcc-scale-factor-1 | 57.87            | 58.92          | 1.81           |


|      test_name      | from_server_name |           from_server_version            | from_tps | to_server_name |            to_server_version             | to_tps | percent_change |
|---------------------|------------------|------------------------------------------|----------|----------------|------------------------------------------|--------|----------------|
| tpcc-scale-factor-1 | dolt             | f4e529af734d3e4b05463d1b64d9d69aaf9f8fa3 | 41.65    | dolt           | e8f4ead63e06724a0239c44a0234a328e197d49f | 41.42  | -0.55          |