@coffeegoddd 2024-11-19
@max-hoffman __DOLT__
 |      read_tests      | from_latency | to_latency | percent_change |
|----------------------|--------------|------------|----------------|
| covering_index_scan  | 0.62         | 0.69       | 11.29          |
| groupby_scan         | 16.71        | 16.41      | -1.8           |
| index_join           | 2.26         | 2.26       | 0.0            |
| index_join_scan      | 1.82         | 1.44       | -20.88         |
| index_scan           | 54.83        | 54.83      | 0.0            |
| oltp_point_select    | 0.27         | 0.27       | 0.0            |
| oltp_read_only       | 5.37         | 5.37       | 0.0            |
| select_random_points | 0.65         | 0.65       | 0.0            |
| select_random_ranges | 0.64         | 0.64       | 0.0            |
| table_scan           | 55.82        | 55.82      | 0.0            |
| types_table_scan     | 144.97       | 142.39     | -1.78          |


|      write_tests      | from_latency | to_latency | percent_change |
|-----------------------|--------------|------------|----------------|
| oltp_delete_insert    | 5.88         | 5.88       | 0.0            |
| oltp_insert           | 2.91         | 2.91       | 0.0            |
| oltp_read_write       | 11.45        | 11.45      | 0.0            |
| oltp_update_index     | 2.97         | 2.97       | 0.0            |
| oltp_update_non_index | 2.91         | 2.91       | 0.0            |
| oltp_write_only       | 5.99         | 5.99       | 0.0            |
| types_delete_insert   | 6.21         | 6.21       | 0.0            |